SONAR-6831 move remaining classes out of package computation
ComputeEngineProcessingModule will be moved in a separate commit which will refactor CE's modules
