+++ server/sonar-server/src/main/java/org/sonar/server/computation/container/ReportComputeEngineContainerPopulator.java
@@ -24,8 +24,8 @@ import java.util.List;
 import org.sonar.core.issue.tracking.Tracker;
 import org.sonar.core.platform.ContainerPopulator;
 import org.sonar.server.computation.queue.CeTask;
-import org.sonar.server.computation.ComputationStepExecutor;
-import org.sonar.server.computation.ComputationTempFolderProvider;
+import org.sonar.server.computation.step.ComputationStepExecutor;
+import org.sonar.server.computation.filesystem.ComputationTempFolderProvider;
 import org.sonar.server.computation.analysis.ReportAnalysisMetadataHolder;
 import org.sonar.server.computation.batch.BatchReportDirectoryHolderImpl;
 import org.sonar.server.computation.batch.BatchReportReaderImpl;
