+++ server/sonar-server/src/main/java/org/sonar/server/computation/property/package-info.java
@@ -17,22 +17,8 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation.queue.report;
 
-import org.sonar.db.ce.CeActivityDto;
-import org.sonar.server.computation.queue.CeQueueListener;
-import org.sonar.server.computation.queue.CeTask;
+@ParametersAreNonnullByDefault
+package org.sonar.server.computation.property;
 
-public class CleanReportQueueListener implements CeQueueListener {
-
-  private final ReportFiles reportFiles;
-
-  public CleanReportQueueListener(ReportFiles reportFiles) {
-    this.reportFiles = reportFiles;
-  }
-
-  @Override
-  public void onRemoved(CeTask task, CeActivityDto.Status status) {
-    reportFiles.deleteIfExists(task.getUuid());
-  }
-}
+import javax.annotation.ParametersAreNonnullByDefault;
