+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/report/ReportFiles.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue.report;
 
 import java.io.File;
 import java.io.InputStream;
