+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/report/ReportSubmitter.java
@@ -30,7 +30,6 @@ import org.sonar.db.ce.CeTaskTypes;
 import org.sonar.db.component.ComponentDto;
 import org.sonar.server.component.ComponentService;
 import org.sonar.server.component.NewComponent;
-import org.sonar.server.computation.ReportFiles;
 import org.sonar.server.computation.queue.CeQueue;
 import org.sonar.server.computation.queue.CeTask;
 import org.sonar.server.computation.queue.CeTaskSubmit;
